requirements.txt
apache-airflow[celery]
apache-airflow[redis]
apache-airflow[rabbitmq]
apache-airflow[cncf.kubernetes]
connexion[swagger-ui]
apache-airflow-providers-papermill
apache-airflow-providers-celery
apache-airflow-providers-redis
psycopg2-binary
ipykernel
requests
pyspark